+++ tests/unit/devices/WH41.spec.ts
@@ -25,6 +25,7 @@ configs.forEach(config => {
     });
 
     beforeEach('Reset config', () => {
+      platform.baseStationInfo.protocol = "Ecowitt";
       platform.config.nameOverrides = [];
       platform.config.hidden = {};
     });